% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\docType{data}
\name{Clusters}
\alias{Clusters}
\title{Data of Hierarchical clustering result on SOM}
\format{
A tibble with 2145 rows and 20 variables, I just pick up the important columns:
\describe{
\item{hc}{Results of hierarchical clustering was recorded. }
\item{\code{subset2020$BFS_nummer}}{It should be used for the further merging with Swiss shapefile. no size}
}
}
\source{
\url{https://applicationspub.unil.ch/interpub/noauth/php/Un/UnPers.php?PerNum=1034786&LanCode=8}
}
\usage{
Clusters
}
\description{
Brief statement: This dataset should be used for the further mapping in R.
}
\keyword{datasets}
