% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGCdata-package.R
\docType{package}
\name{SGCdata-package}
\alias{SGCdata}
\alias{SGCdata-package}
\title{SGCdata: Data packages with courses lab dataset}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Haokun Liu \email{haokun.liu@unil.ch}

Authors:
\itemize{
  \item Marj Tonini \email{marj.tonini@unil.ch}
}

}
\keyword{internal}
